SPARKC-391: Fix conversion of LocalDate to Joda LocalDate
We transform LocalDate objects to JodaLocal date objects so that we'll
have a serializable AnyRef fo moving around in Spark. Previously this
was done by just using the default constructor which assumes the
timestamp provided is in the DefaultTimezone. This is incorrect at
certain times of the year in certain places since LocalDate will give
back the Timestamp in GMT.

To fix this we always pass GMT as the Timezone to convert LocalDates
with.