SPARKC-391: Delagate all DateTime Parsing to JodaTime
Previously we were using some of our own home grown normalization of
timestamps for converting too and from java.sql.Date. This has led to
some errors and difficult to maintain code. In this commit we replace
all previous methods with a pass through of Joda's internal normalizing
machinery.

In addition we now run through all our Date tests in both PST, and CET
on Jenkins